refactor(settings): reorganize into collapsible groups with shared renderer
- Extract shared preferences renderer (preferences-content.ts) for reuse
  by both web modal and desktop settings window
- Replace overflowing General tab (11 flat sections) with 5 collapsible
  <details> groups: Display, Intelligence, Media, Panels, Data & Community
- Remove redundant Status tab (StatusPanel already exists as standalone)
- Remove statusPanel/isGlobeMode/onMapModeChange from UnifiedSettingsConfig
- Rename "General" tab to "Settings" with new tabSettings i18n key
- Event listeners use AbortController for clean lifecycle management
- Add preferences.* i18n keys to all 21 locales with proper translations
- UnifiedSettings.ts: 838 → 428 lines (-49%)
