+++ src/locales/bg.json
@@ -140,6 +140,7 @@
     "toggleTheme": "Превключване на тъмен/светъл режим",
     "panelDisplayCaption": "Изберете кои панели да се показват на табло",
     "tabGeneral": "Основно",
+    "tabSettings": "Настройки",
     "tabPanels": "Панели",
     "tabSources": "Източници",
     "languageLabel": "Език",
@@ -2335,5 +2336,23 @@
     "retry": "Опитайте отново",
     "refresh": "Обновяване",
     "all": "Всички"
+  },
+  "preferences": {
+    "display": "Дисплей",
+    "intelligence": "Интелигентност",
+    "media": "Медия",
+    "panels": "Панели",
+    "dataAndCommunity": "Данни и общност",
+    "theme": "Тема",
+    "themeDesc": "Автоматично следва системните настройки.",
+    "themeAuto": "Автоматично (следвай системата)",
+    "themeDark": "Тъмна",
+    "themeLight": "Светла",
+    "mapProvider": "Доставчик на карти",
+    "mapProviderDesc": "Изберете откъде да се зареждат картните плочки.",
+    "mapTheme": "Тема на картата",
+    "mapThemeDesc": "Визуален стил на картните плочки.",
+    "globePreset": "Визуален предварителен набор",
+    "globePresetDesc": "Превключване между класически и подобрени визуализации."
   }
 }
