+++ src/locales/el.json
@@ -140,6 +140,7 @@
     "toggleTheme": "Εναλλαγή σκοτεινής/φωτεινής λειτουργίας",
     "panelDisplayCaption": "Επιλέξτε ποια πάνελ θα εμφανίζονται στον πίνακα ελέγχου",
     "tabGeneral": "Γενικά",
+    "tabSettings": "Ρυθμίσεις",
     "tabPanels": "Πάνελ",
     "tabSources": "Πηγές",
     "languageLabel": "Γλώσσα",
@@ -2335,5 +2336,23 @@
     "retry": "Επανάληψη",
     "refresh": "Ανανέωση",
     "all": "Όλα"
+  },
+  "preferences": {
+    "display": "Οθόνη",
+    "intelligence": "Νοημοσύνη",
+    "media": "Μέσα",
+    "panels": "Πίνακες",
+    "dataAndCommunity": "Δεδομένα & Κοινότητα",
+    "theme": "Θέμα",
+    "themeDesc": "Αυτόματα ακολουθεί τις ρυθμίσεις συστήματος.",
+    "themeAuto": "Αυτόματο (ακολούθηση συστήματος)",
+    "themeDark": "Σκοτεινό",
+    "themeLight": "Φωτεινό",
+    "mapProvider": "Πάροχος πλακιδίων χάρτη",
+    "mapProviderDesc": "Επιλέξτε πηγή φόρτωσης πλακιδίων χάρτη.",
+    "mapTheme": "Θέμα χάρτη",
+    "mapThemeDesc": "Οπτικό στυλ των πλακιδίων χάρτη.",
+    "globePreset": "Οπτική προεπιλογή",
+    "globePresetDesc": "Εναλλαγή μεταξύ κλασικής και βελτιωμένης απεικόνισης."
   }
 }
