+++ src/locales/ko.json
@@ -140,6 +140,7 @@
     "toggleTheme": "다크/라이트 모드 전환",
     "panelDisplayCaption": "대시보드에 표시할 패널을 선택하세요",
     "tabGeneral": "일반",
+    "tabSettings": "설정",
     "tabPanels": "패널",
     "tabSources": "소스",
     "languageLabel": "언어",
@@ -2335,5 +2336,23 @@
     "retry": "재시도",
     "refresh": "새로고침",
     "all": "전체"
+  },
+  "preferences": {
+    "display": "디스플레이",
+    "intelligence": "인텔리전스",
+    "media": "미디어",
+    "panels": "패널",
+    "dataAndCommunity": "데이터 및 커뮤니티",
+    "theme": "테마",
+    "themeDesc": "자동으로 시스템 설정을 따릅니다.",
+    "themeAuto": "자동 (시스템 따르기)",
+    "themeDark": "다크",
+    "themeLight": "라이트",
+    "mapProvider": "지도 타일 제공자",
+    "mapProviderDesc": "지도 타일을 로드할 위치를 선택하세요.",
+    "mapTheme": "지도 테마",
+    "mapThemeDesc": "지도 타일의 시각적 스타일.",
+    "globePreset": "시각적 프리셋",
+    "globePresetDesc": "클래식과 향상된 글로브 시각화 간 전환."
   }
 }
