+++ src/locales/zh.json
@@ -140,6 +140,7 @@
     "toggleTheme": "切换深色/浅色模式",
     "panelDisplayCaption": "选择要在仪表板上显示的面板",
     "tabGeneral": "常规",
+    "tabSettings": "设置",
     "tabPanels": "面板",
     "tabSources": "来源",
     "languageLabel": "语言",
@@ -2335,5 +2336,23 @@
     "retry": "Retry",
     "refresh": "Refresh",
     "all": "全部"
+  },
+  "preferences": {
+    "display": "显示",
+    "intelligence": "智能",
+    "media": "媒体",
+    "panels": "面板",
+    "dataAndCommunity": "数据与社区",
+    "theme": "主题",
+    "themeDesc": "自动跟随系统偏好。",
+    "themeAuto": "自动（跟随系统）",
+    "themeDark": "深色",
+    "themeLight": "浅色",
+    "mapProvider": "地图图块提供者",
+    "mapProviderDesc": "选择地图图块的加载来源。",
+    "mapTheme": "地图主题",
+    "mapThemeDesc": "地图图块的视觉风格。",
+    "globePreset": "视觉预设",
+    "globePresetDesc": "在经典和增强的地球视觉之间切换。"
   }
 }
