+++ src/services/preferences-content.ts
@@ -1,0 +1,348 @@
+import { LANGUAGES, getCurrentLanguage, changeLanguage, t } from '@/services/i18n';
+import { getAiFlowSettings, setAiFlowSetting, getStreamQuality, setStreamQuality, STREAM_QUALITY_OPTIONS } from '@/services/ai-flow-settings';
+import { getMapProvider, setMapProvider, MAP_PROVIDER_OPTIONS, MAP_THEME_OPTIONS, getMapTheme, setMapTheme, type MapProvider } from '@/config/basemap';
+import { getLiveStreamsAlwaysOn, setLiveStreamsAlwaysOn } from '@/services/live-stream-settings';
+import { getGlobeVisualPreset, setGlobeVisualPreset, GLOBE_VISUAL_PRESET_OPTIONS, type GlobeVisualPreset } from '@/services/globe-render-settings';
+import type { StreamQuality } from '@/services/ai-flow-settings';
+import { getThemePreference, setThemePreference, type ThemePreference } from '@/utils/theme-manager';
+import { escapeHtml } from '@/utils/sanitize';
+import { trackLanguageChange } from '@/services/analytics';
+import { exportSettings, importSettings, type ImportResult } from '@/utils/settings-persistence';
+
+const DESKTOP_RELEASES_URL = 'https://github.com/koala73/worldmonitor/releases';
+
+export interface PreferencesHost {
+  isDesktopApp: boolean;
+  onMapProviderChange?: (provider: MapProvider) => void;
+}
+
+export interface PreferencesResult {
+  html: string;
+  attach: (container: HTMLElement) => () => void;
+}
+
+function toggleRowHtml(id: string, label: string, desc: string, checked: boolean): string {
+  return `
+    <div class="ai-flow-toggle-row">
+      <div class="ai-flow-toggle-label-wrap">
+        <div class="ai-flow-toggle-label">${label}</div>
+        <div class="ai-flow-toggle-desc">${desc}</div>
+      </div>
+      <label class="ai-flow-switch">
+        <input type="checkbox" id="${id}"${checked ? ' checked' : ''}>
+        <span class="ai-flow-slider"></span>
+      </label>
+    </div>
+  `;
+}
+
+function renderMapThemeDropdown(container: HTMLElement, provider: MapProvider): void {
+  const select = container.querySelector<HTMLSelectElement>('#us-map-theme');
+  if (!select) return;
+  const currentTheme = getMapTheme(provider);
+  select.innerHTML = MAP_THEME_OPTIONS[provider]
+    .map(opt => `<option value="${opt.value}"${opt.value === currentTheme ? ' selected' : ''}>${escapeHtml(opt.label)}</option>`)
+    .join('');
+}
+
+function updateAiStatus(container: HTMLElement): void {
+  const settings = getAiFlowSettings();
+  const dot = container.querySelector('#usStatusDot');
+  const text = container.querySelector('#usStatusText');
+  if (!dot || !text) return;
+
+  dot.className = 'ai-flow-status-dot';
+  if (settings.cloudLlm && settings.browserModel) {
+    dot.classList.add('active');
+    text.textContent = t('components.insights.aiFlowStatusCloudAndBrowser');
+  } else if (settings.cloudLlm) {
+    dot.classList.add('active');
+    text.textContent = t('components.insights.aiFlowStatusActive');
+  } else if (settings.browserModel) {
+    dot.classList.add('browser-only');
+    text.textContent = t('components.insights.aiFlowStatusBrowserOnly');
+  } else {
+    dot.classList.add('disabled');
+    text.textContent = t('components.insights.aiFlowStatusDisabled');
+  }
+}
+
+export function renderPreferences(host: PreferencesHost): PreferencesResult {
+  const settings = getAiFlowSettings();
+  const currentLang = getCurrentLanguage();
+  let html = '';
+
+  // ── Display group ──
+  html += `<details class="wm-pref-group" open>`;
+  html += `<summary>${t('preferences.display')}</summary>`;
+  html += `<div class="wm-pref-group-content">`;
+
+  // Appearance
+  const currentThemePref = getThemePreference();
+  html += `<div class="ai-flow-toggle-row">
+    <div class="ai-flow-toggle-label-wrap">
+      <div class="ai-flow-toggle-label">${t('preferences.theme')}</div>
+      <div class="ai-flow-toggle-desc">${t('preferences.themeDesc')}</div>
+    </div>
+  </div>`;
+  html += `<select class="unified-settings-select" id="us-theme">`;
+  for (const opt of [
+    { value: 'auto', label: t('preferences.themeAuto') },
+    { value: 'dark', label: t('preferences.themeDark') },
+    { value: 'light', label: t('preferences.themeLight') },
+  ] as { value: ThemePreference; label: string }[]) {
+    const selected = opt.value === currentThemePref ? ' selected' : '';
+    html += `<option value="${opt.value}"${selected}>${escapeHtml(opt.label)}</option>`;
+  }
+  html += `</select>`;
+
+  // Map tile provider
+  const currentProvider = getMapProvider();
+  html += `<div class="ai-flow-toggle-row">
+    <div class="ai-flow-toggle-label-wrap">
+      <div class="ai-flow-toggle-label">${t('preferences.mapProvider')}</div>
+      <div class="ai-flow-toggle-desc">${t('preferences.mapProviderDesc')}</div>
+    </div>
+  </div>`;
+  html += `<select class="unified-settings-select" id="us-map-provider">`;
+  for (const opt of MAP_PROVIDER_OPTIONS) {
+    const selected = opt.value === currentProvider ? ' selected' : '';
+    html += `<option value="${opt.value}"${selected}>${escapeHtml(opt.label)}</option>`;
+  }
+  html += `</select>`;
+
+  // Map theme
+  const currentMapTheme = getMapTheme(currentProvider);
+  html += `<div class="ai-flow-toggle-row">
+    <div class="ai-flow-toggle-label-wrap">
+      <div class="ai-flow-toggle-label">${t('preferences.mapTheme')}</div>
+      <div class="ai-flow-toggle-desc">${t('preferences.mapThemeDesc')}</div>
+    </div>
+  </div>`;
+  html += `<select class="unified-settings-select" id="us-map-theme">`;
+  for (const opt of MAP_THEME_OPTIONS[currentProvider]) {
+    const selected = opt.value === currentMapTheme ? ' selected' : '';
+    html += `<option value="${opt.value}"${selected}>${escapeHtml(opt.label)}</option>`;
+  }
+  html += `</select>`;
+
+  html += toggleRowHtml('us-map-flash', t('components.insights.mapFlashLabel'), t('components.insights.mapFlashDesc'), settings.mapNewsFlash);
+
+  // 3D Globe Visual Preset
+  const currentPreset = getGlobeVisualPreset();
+  html += `<div class="ai-flow-toggle-row">
+    <div class="ai-flow-toggle-label-wrap">
+      <div class="ai-flow-toggle-label">${t('preferences.globePreset')}</div>
+      <div class="ai-flow-toggle-desc">${t('preferences.globePresetDesc')}</div>
+    </div>
+  </div>`;
+  html += `<select class="unified-settings-select" id="us-globe-visual-preset">`;
+  for (const opt of GLOBE_VISUAL_PRESET_OPTIONS) {
+    const selected = opt.value === currentPreset ? ' selected' : '';
+    html += `<option value="${opt.value}"${selected}>${escapeHtml(opt.label)}</option>`;
+  }
+  html += `</select>`;
+
+  // Language
+  html += `<div class="ai-flow-section-label">${t('header.languageLabel')}</div>`;
+  html += `<select class="unified-settings-lang-select" id="us-language">`;
+  for (const lang of LANGUAGES) {
+    const selected = lang.code === currentLang ? ' selected' : '';
+    html += `<option value="${lang.code}"${selected}>${lang.flag} ${escapeHtml(lang.label)}</option>`;
+  }
+  html += `</select>`;
+  if (currentLang === 'vi') {
+    html += `<div class="ai-flow-toggle-desc">${t('components.languageSelector.mapLabelsFallbackVi')}</div>`;
+  }
+
+  html += `</div></details>`;
+
+  // ── Intelligence group ──
+  html += `<details class="wm-pref-group">`;
+  html += `<summary>${t('preferences.intelligence')}</summary>`;
+  html += `<div class="wm-pref-group-content">`;
+
+  if (!host.isDesktopApp) {
+    html += toggleRowHtml('us-cloud', t('components.insights.aiFlowCloudLabel'), t('components.insights.aiFlowCloudDesc'), settings.cloudLlm);
+    html += toggleRowHtml('us-browser', t('components.insights.aiFlowBrowserLabel'), t('components.insights.aiFlowBrowserDesc'), settings.browserModel);
+    html += `<div class="ai-flow-toggle-warn" style="display:${settings.browserModel ? 'block' : 'none'}">${t('components.insights.aiFlowBrowserWarn')}</div>`;
+    html += `
+      <div class="ai-flow-cta">
+        <div class="ai-flow-cta-title">${t('components.insights.aiFlowOllamaCta')}</div>
+        <div class="ai-flow-cta-desc">${t('components.insights.aiFlowOllamaCtaDesc')}</div>
+        <a href="${DESKTOP_RELEASES_URL}" target="_blank" rel="noopener noreferrer" class="ai-flow-cta-link">${t('components.insights.aiFlowDownloadDesktop')}</a>
+      </div>
+    `;
+  }
+
+  html += toggleRowHtml('us-headline-memory', t('components.insights.headlineMemoryLabel'), t('components.insights.headlineMemoryDesc'), settings.headlineMemory);
+
+  html += `</div></details>`;
+
+  // ── Media group ──
+  html += `<details class="wm-pref-group">`;
+  html += `<summary>${t('preferences.media')}</summary>`;
+  html += `<div class="wm-pref-group-content">`;
+
+  const currentQuality = getStreamQuality();
+  html += `<div class="ai-flow-toggle-row">
+    <div class="ai-flow-toggle-label-wrap">
+      <div class="ai-flow-toggle-label">${t('components.insights.streamQualityLabel')}</div>
+      <div class="ai-flow-toggle-desc">${t('components.insights.streamQualityDesc')}</div>
+    </div>
+  </div>`;
+  html += `<select class="unified-settings-select" id="us-stream-quality">`;
+  for (const opt of STREAM_QUALITY_OPTIONS) {
+    const selected = opt.value === currentQuality ? ' selected' : '';
+    html += `<option value="${opt.value}"${selected}>${escapeHtml(opt.label)}</option>`;
+  }
+  html += `</select>`;
+
+  html += toggleRowHtml(
+    'us-live-streams-always-on',
+    t('components.insights.streamAlwaysOnLabel'),
+    t('components.insights.streamAlwaysOnDesc'),
+    getLiveStreamsAlwaysOn(),
+  );
+
+  html += `</div></details>`;
+
+  // ── Panels group ──
+  html += `<details class="wm-pref-group">`;
+  html += `<summary>${t('preferences.panels')}</summary>`;
+  html += `<div class="wm-pref-group-content">`;
+  html += toggleRowHtml('us-badge-anim', t('components.insights.badgeAnimLabel'), t('components.insights.badgeAnimDesc'), settings.badgeAnimation);
+  html += `</div></details>`;
+
+  // ── Data & Community group ──
+  html += `<details class="wm-pref-group">`;
+  html += `<summary>${t('preferences.dataAndCommunity')}</summary>`;
+  html += `<div class="wm-pref-group-content">`;
+  html += `
+    <div class="us-data-mgmt">
+      <button type="button" class="settings-btn settings-btn-secondary" id="usExportBtn">${t('components.settings.exportSettings')}</button>
+      <button type="button" class="settings-btn settings-btn-secondary" id="usImportBtn">${t('components.settings.importSettings')}</button>
+      <input type="file" id="usImportInput" accept=".json" class="us-hidden-input" />
+    </div>
+    <div class="us-data-mgmt-toast" id="usDataMgmtToast"></div>
+  `;
+  html += `<a href="https://github.com/koala73/worldmonitor/discussions/94" target="_blank" rel="noopener noreferrer" class="us-discussion-link">
+    <span class="us-discussion-dot"></span>
+    <span>${t('components.community.joinDiscussion')}</span>
+  </a>`;
+  html += `</div></details>`;
+
+  // AI status footer (web-only)
+  if (!host.isDesktopApp) {
+    html += `<div class="ai-flow-popup-footer"><span class="ai-flow-status-dot" id="usStatusDot"></span><span class="ai-flow-status-text" id="usStatusText"></span></div>`;
+  }
+
+  return {
+    html,
+    attach(container: HTMLElement): () => void {
+      const ac = new AbortController();
+      const { signal } = ac;
+
+      container.addEventListener('change', (e) => {
+        const target = e.target as HTMLInputElement;
+
+        if (target.id === 'usImportInput') {
+          const file = target.files?.[0];
+          if (!file) return;
+          importSettings(file).then((result: ImportResult) => {
+            showToast(container, t('components.settings.importSuccess', { count: String(result.keysImported) }), true);
+          }).catch(() => {
+            showToast(container, t('components.settings.importFailed'), false);
+          });
+          target.value = '';
+          return;
+        }
+
+        if (target.id === 'us-stream-quality') {
+          setStreamQuality(target.value as StreamQuality);
+          return;
+        }
+        if (target.id === 'us-globe-visual-preset') {
+          setGlobeVisualPreset(target.value as GlobeVisualPreset);
+          return;
+        }
+        if (target.id === 'us-theme') {
+          setThemePreference(target.value as ThemePreference);
+          return;
+        }
+        if (target.id === 'us-map-provider') {
+          const provider = target.value as MapProvider;
+          setMapProvider(provider);
+          renderMapThemeDropdown(container, provider);
+          host.onMapProviderChange?.(provider);
+          window.dispatchEvent(new CustomEvent('map-theme-changed'));
+          return;
+        }
+        if (target.id === 'us-map-theme') {
+          const provider = getMapProvider();
+          setMapTheme(provider, target.value);
+          window.dispatchEvent(new CustomEvent('map-theme-changed'));
+          return;
+        }
+        if (target.id === 'us-live-streams-always-on') {
+          setLiveStreamsAlwaysOn(target.checked);
+          return;
+        }
+        if (target.id === 'us-language') {
+          trackLanguageChange(target.value);
+          void changeLanguage(target.value);
+          return;
+        }
+        if (target.id === 'us-cloud') {
+          setAiFlowSetting('cloudLlm', target.checked);
+          updateAiStatus(container);
+        } else if (target.id === 'us-browser') {
+          setAiFlowSetting('browserModel', target.checked);
+          const warn = container.querySelector('.ai-flow-toggle-warn') as HTMLElement;
+          if (warn) warn.style.display = target.checked ? 'block' : 'none';
+          updateAiStatus(container);
+        } else if (target.id === 'us-map-flash') {
+          setAiFlowSetting('mapNewsFlash', target.checked);
+        } else if (target.id === 'us-headline-memory') {
+          setAiFlowSetting('headlineMemory', target.checked);
+        } else if (target.id === 'us-badge-anim') {
+          setAiFlowSetting('badgeAnimation', target.checked);
+        }
+      }, { signal });
+
+      container.addEventListener('click', (e) => {
+        const target = e.target as HTMLElement;
+        if (target.closest('#usExportBtn')) {
+          try {
+            exportSettings();
+            showToast(container, t('components.settings.exportSuccess'), true);
+          } catch {
+            showToast(container, t('components.settings.exportFailed'), false);
+          }
+          return;
+        }
+        if (target.closest('#usImportBtn')) {
+          container.querySelector<HTMLInputElement>('#usImportInput')?.click();
+          return;
+        }
+      }, { signal });
+
+      if (!host.isDesktopApp) updateAiStatus(container);
+
+      return () => ac.abort();
+    },
+  };
+}
+
+function showToast(container: HTMLElement, msg: string, success: boolean): void {
+  const toast = container.querySelector('#usDataMgmtToast');
+  if (!toast) return;
+  toast.className = `us-data-mgmt-toast ${success ? 'ok' : 'error'}`;
+  toast.innerHTML = success
+    ? `${escapeHtml(msg)} <a href="#" class="us-toast-reload">${t('components.settings.reloadNow')}</a>`
+    : escapeHtml(msg);
+  toast.querySelector('.us-toast-reload')?.addEventListener('click', (e) => {
+    e.preventDefault();
+    window.location.reload();
+  });
+}
